+++ gen/com/intellij/idea/plugin/hybris/impex/ImpexParser.java
@@ -1578,10 +1578,11 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   /* ********************************************************** */
   // FIELD_VALUE
   //     | FIELD_VALUE_JAR_PREFIX
+  //     | FIELD_VALUE_EXPLODED_JAR_PREFIX
   //     | FIELD_VALUE_FILE_PREFIX
   //     | FIELD_VALUE_ZIP_PREFIX
   //     | FIELD_VALUE_HTTP_PREFIX
-  //     | FIELD_VALUE_URL
+  // //    | FIELD_VALUE_URL
   //     | BOOLEAN
   //     | DIGIT
   //     | string
@@ -1598,10 +1599,10 @@ public class ImpexParser implements PsiParser, LightPsiParser {
     boolean r;
     r = consumeToken(b, FIELD_VALUE);
     if (!r) r = consumeToken(b, FIELD_VALUE_JAR_PREFIX);
+    if (!r) r = consumeToken(b, FIELD_VALUE_EXPLODED_JAR_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_FILE_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_ZIP_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_HTTP_PREFIX);
-    if (!r) r = consumeToken(b, FIELD_VALUE_URL);
     if (!r) r = consumeToken(b, BOOLEAN);
     if (!r) r = consumeToken(b, DIGIT);
     if (!r) r = string(b, l + 1);